Fix scripts
* Added README.md for mc;
* Update TODO about deno error on clean installation yt-dlp.

+++ install_scripts/yt-dlp.sh
@@ -6,7 +6,12 @@
 set -Eeuo pipefail
 
 source "$MYSCRIPTS/utils.sh"
-check_dependencies apt curl deno ffmpeg ffprobe
 
-sudo curl -L https://github.com/yt-dlp/yt-dlp/releases/latest/download/yt-dlp -o /usr/local/bin/yt-dlp
-sudo chmod a+rx /usr/local/bin/yt-dlp  # Make executable
+echo
+read -r -p "Install yt-dlp? [y/N] " response
+if [[ "$response" =~ ^([yY][eE][sS]|[yY])$ ]]
+then
+  check_dependencies apt curl deno ffmpeg ffprobe
+  sudo curl -L https://github.com/yt-dlp/yt-dlp/releases/latest/download/yt-dlp -o /usr/local/bin/yt-dlp
+  sudo chmod a+rx /usr/local/bin/yt-dlp  # Make executable
+fi
